requirements.txt
fastapi
uvicorn
yfinance
pandas
numpy
scikit-learn
torch
joblib
requests
python-dotenv
newspaper3k
lxml_html_clean
textblob
